Behave correctly with 2 argument + domain scenario.
Previously all errors were bubbled up to the domain even if a specific
error callback was provided, which prevented any custom error handling
even for expected errors.  Now if two callbacks are given, one for error
and one for success, even if a domain is active the error callback will
be called.  This brings parity to the behavior of okay when domains are
not active, and frankly the other way was broken.